Correction de warnings (casting)

+++ src/Core/Timer.h
@@ -8,10 +8,10 @@ public:
 	Timer();
 	~Timer();
 
-	int GetStartTime() { return startTime; }
-	int GetCurrTime() { return currTime; }
-	int GetTime() { return (currTime - startTime); }
-	int GetTimeInSeconds() { return (currTime - startTime) / 1000; }
+	int64_t GetStartTime() { return startTime; }
+	int64_t GetCurrTime() { return currTime; }
+	int64_t GetTime() { return (currTime - startTime); }
+	int64_t GetTimeInSeconds() { return (currTime - startTime) / 1000; }
 	Uint64 GetFCurrTime() { return fCurrTime; }
 	Uint64 GetFPrevTime() { return fPrevTime; }
 	double GetDeltaTime() { return deltaTime; }
